back_end/Automatic_classification_of_news_articles_in_Arabic/requirements.txt
ArabicOcr==1.1.6
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
easyocr==1.7.1
filelock==3.14.0
fsspec==2024.3.1
imageio==2.34.1
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.2
lazy_loader==0.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
pyclipper==1.3.0.post5
pytesseract==0.3.10
python-bidi==0.4.2
PyYAML==6.0.1
regex==2024.4.28
scikit-image==0.23.2
scikit-learn==1.4.1.post1
scipy==1.13.0
shapely==2.0.4
six==1.16.0
sqlparse==0.5.0
sympy==1.12
tbb==2021.12.0
threadpoolctl==3.5.0
tifffile==2024.5.3
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
